feat: Added `jdk default`, `jdk home` and `jdk java-env` commands
These commands were added to facilitate the use of the Jbang's JDKs
by the user. This way it's not necessary to install additional JDKs
if we want to use Java and have previously let Jbang install a JDK.
`jdk default` set the given JDK as "current".
`jdk home` returns the path of the current or the given JDK.
`jdk java-env` returns the shell commands to make the current or
the given JDK usable in the user's shell.