refactor: Now using special exit code to allow printing to stdout
Before we would either print as much output to stderr, which is not
in line with the Rule of Least Surprise, or we had to print the output
as part of a shell command like `echo` or `cat`. Now we have a special
exit code that will tell the startup scripts to either execute Jbang's
output or to simply display it.