Jar output via jbang export (#565)
* refactor: turn Script into a stateless data object

* fix: Correct format of (Boot-)Class-Path items in JAR manifests

Was using the same format as the java(c) command on the CLI but
class paths in a MANIFEST files are actually URLs (without the
protocol bit). So names with spaces are not handled by quoting
but by url-encoding.

* feat: Always add all dependencies to the JAR manifest

This allows (most) JARs to be run directly by using `jar -jar`.

* feat: Added `--output` option to `build`

Allows the user to get a copy of the application JAR at a location
of their choice.

* refactor: Moved `AliasUtil.getCwd()` to `Util`

* feat: Introduced `export` command

* refactor: Moved CLI option `userParams` from base class to `Run`