feat: Add `export --portable`
A fairly naive export mechanism to have an option on export
that says instead of just copying the built jar into a directory; make it a
slim jar with a `libs` directory next to it.

the portable jar has all its classpath dependencies updated to point to libs/* instead
with relative paths letting you move it anywhere to just run.

+ turned of annoying console logging in karate

Fixes #619