feat: Made sure compiled JARs are used more and still valid
Before compiled JARs wouldn't be rebuilt but the source file would
still be parsed for all kinds of meta information (like dependencies).
This makes sure that whenver possible only the JAR will be used.
It also checks if the JAR's dependencies are still available and
if not it will rebuild.