feat: We now check if dependencies are still up-to-date
File timestamps of all dependencies are now stored in the dependency
cache file and are checked against the actual files runtime. If they
don't coincide the script will be rebuilt first.
We now also check the dependencies stored within a JAR file's MANIFEST
and rebuild the script if they are missing or out-of-date.