refactor: Do more efficient Catalog lookups
Before we would always create a complete "merged" catalog containing
all the infomration from all the available files in the current
context (local catalogs recursively going up the filesystem,
user catalog and implicit catalog). Then we'd see if the information
we were looking for could be foiund within. But this is ineffecient
if the information happened to be very "near". So the new code just
looks into the catalogs one by one and returns the information
immediately once it's found.