chore: enable build and promotion of snapcraft
This adds tag update of jbang-snap which will trigger
a build on snapcraft visible at https://snapcraft.io/jbang/builds

Then we wait ~20 minutes before we do a scripted promote
of every builds with the same version as we tagged with.

We do it this way as snapcraft for reasons I just don't grok
do not want to enable automated releases..so we have to work our
way around it.