fix: Now correctly handles out-of-date dependencies
Updating the dependency cache now correctly only writes out the
information that was actually retrieved and updated. Before the
code would always write out the latest timestamps, updating them
for *all* files in the cache, even the ones we weren't even
retrieving. Because resolving dependencies is a multi-step
process this basically ruined the information that was needed
for the following steps.

Fixes #757