fix: Update should now work on Windows
When a new update gets downloaded we won't try to overwrite
the jbang.jar on Windows because that would fail anyway.
Instead we copy the new jar next to the old jar as `jbang.jar.new`
and we overwrite the old version when the user runs jbang the
next time.

Fixes #823