fix: Correct use of @-files (#884)
* fix: Correct use of @-files

We now correctly quote the first CLI argument as well as all the
arguments in the @-file. We also take into account now the values
of //JAVA and --java when determining which Java version we're
running.

Fixes #874

* fix: No longer unnecessarily recompiles

The check to determine if a script had to be recompiled
because of a mismatch between the JDK version that was
used to compile the JAR and the available JDK was not
entirely correct and was causing unnecessary recompiles.