Better deps handling (#1111)
* refactor: Improved dependency handling

`Source` dependency handling has been improved so that regardless
of the sources or commind line options like `--deps` we'll only
do a single call to the dependency resolver.

* refactor: Aether now resolves all deps in a single go

This fixes any problems with duplicate and conflicting dependencies
because we used to handle each dependency separately. The downside,
for now, is that we can't log the dependency resolution to the
output for the user to see what's happening.