feat: Allow jbang to install itself
Normally when you run `jbang app install jbang` a fresh copy of the
latest version of Jbang will be downloaded, unpacked and installed.
But this patch allows the currently running jbang to be used as the
source for the files to install. This is done by supplying the
option `--offline` to prevent Jbang from downloading the latest version.

See #1122