feat: quality-of-life improvements for `edit`
In case no managed VSCodium is installed, Jbang will now search the
user's PATH for well-known editors and present them as additional
options next to offer of downloading VSCodium.
If the user selects one of the editors found on the PATH, Jbang will
show the `config` command the user can use to set that option as the
default for future invocations.
A new option was added to explicitely exit without doing anything.

Fixes #1210