feat: Added `--code` option for running literal scripts
The `--code` (or `-c`) option either takes a String argument (using `=`)
that will be taken as a literal script or it can appear without argument
in which case the `scriptOrFile` argument will be assumed to contain a
literal script. In both cases the literal script will be executed in
the same way as if the code was passed on the stdin (that is, the code
will be assumed to be Jshell code unless it obviously is plain Java).
Also allows the `--interactive` option to be used without any script
reference at all.
The shortcut `-i` was added for `--interactive`.

Fixes #1242