feat: Added `--code` option for running literal scripts (#1243)
* feat: Added `--code` option for running literal scripts

The `--code` (or `-c`) option either takes a String argument (using `=`)
that will be taken as a literal script or it can appear without argument
in which case the `scriptOrFile` argument will be assumed to contain a
literal script. In both cases the literal script will be executed in
the same way as if the code was passed on the stdin (that is, the code
will be assumed to be Jshell code unless it obviously is plain Java).
Also allows the `--interactive` option to be used without any script
reference at all.
The shortcut `-i` was added for `--interactive`.

Fixes #1242

* docs: Added docs for `--code` code_option

See #1242

* test: Added tests for `--code` option

See #1242

* fix: fixes handling of `//SOURCES` in ephemeral scripts

This is because Jbang tries to look up those paths relative to the
original script location which in the case of scripts that come from
stdin or from the `--code` option is not available. In those cases we
fall back to using the current working directory.

See #1242

* feat: added `--sources` (`-s`) option for adding additional sources

* test: added test for `--sources` option