feat: Support colors under Windows
For various combinations of Windows terminal emulators and shells (e.g. Git Bash) the help displayed by picocli when running `jbang` is garbled with ANSI codes to the point that the help almost can't be deciphered.

According to the [documentation of picocli](https://picocli.info/#_windows) the ANSI support can be enabled by including Jansi and calling `AnsiConsole.systemInstall()`. That is what is being done in this commit.

Note that the Jansi JAR is 215 KB in size.