refactor: major class hierarchy reorg
No code was added in this refactor but the hierarchy was changed
extensively. First what was before called `Source` is now called `Input`
with the only implementing classes `Jar` and `SourceSet`.
The `ScriptSource` and child classes are now simply named `Script` and
don't inherit from `Source` anymore.
And Java isn't the default anymore for `Script`, but a separate
`JavaScript` (yes yes) was created and `Script` is now abstract.