refactor: introduced new builders
Instead of a single `Builder` implementation (`JarBuilder`) we now use
a base implementation and several implementations, one for each of the
support programming languages. `SourceSet`s now know how they should
be built and can return the correct `Builder` for their sources.
A bunch of code was moved from the different `Script` classes to the
new `Builder` classes.