refactor: Big reorg for handling sources (#1272)
* refactor: introduced `Builder` and `JarBuilder`

* refactor: introduced `CmdGenerator` and `DefaultCmdGenerator`

* fix: no longer generating too long commands for PowerShell

Seems that PowerShell is still bound by some pretty limmited command
lines buffers. So we'll just apply the same rules for PowerShell as we
do for CMD.

Fixes #1253

* refactor: move code out of `ScriptSource`

The code for (recursively) obtaining all script information was all
centered on `ScriptSource` which made it hard to create/add alternative
sources of information. This is the first step on the way to remedy that.

* refactor: introduced `SourceSet`

* refactor: major class hierarchy reorg

No code was added in this refactor but the hierarchy was changed
extensively. First what was before called `Source` is now called `Input`
with the only implementing classes `Jar` and `SourceSet`.
The `ScriptSource` and child classes are now simply named `Script` and
don't inherit from `Source` anymore.
And Java isn't the default anymore for `Script`, but a separate
`JavaScript` (yes yes) was created and `Script` is now abstract.

* refactor: introduced new builders

Instead of a single `Builder` implementation (`JarBuilder`) we now use
a base implementation and several implementations, one for each of the
support programming languages. `SourceSet`s now know how they should
be built and can return the correct `Builder` for their sources.
A bunch of code was moved from the different `Script` classes to the
new `Builder` classes.

* refactor: renamed `Script` to `Source`

Now that `Source` wasn't used anymore we could use it again for the
classes that represent source files.

* refactor: renamed `Input` to `Code`

`Input` wasn't really a good name because when a `SourceSet` gets
built it's turned into a `Jar` which is an ouput, not an input.
While both a `SourceSet` and a `Jar` represent code, so `Code` seems
like a good generic name.

* refactor: introduced new CmdGenerators

The `DefaultCmdGenerator` was split into a `BaseCmdGenerator` and pair
of implementations for running jars and jshell scripts.

* fix: fixed CDS handling

* fix: fixed some tests on Windows