feat: fall back to TTY when asking for input on Linux/Mac
This means that the GUI that was introduced a while back to be able to
ask the user for feedback when the stdin was redirected will no longer
be used on Linux and Mac, but only on Windows.
The main advantage of this is that zero-install using `curl` will
work when no GUI is available (eg when using containers).