fix: make aliases work with --repos (#1372)
The "repositories" item from an Alias would not actually be applied
because at the time the Alias would be detected the dependecy resolver
was already created. It had to be made more dynamic so the resolver
can change depending if there's an Alias or not.

Additionally the `--repos` argument for `alias add` would also not get
applied so trying to add something that depended on that repository
being active would fail. This is now also fixed.

Fixes #1371