fix: `wrapper install` now also copies jbang.ps1 (#1391)
We didn't copy this before because having just `jbang.cmd` was thought
to be enough for Windows. But some time ago we actually removed all
installation code from the batch file and delegate any installation
requirements to the .ps1 file. So the wrapper installation was missing
a file on Windows.

Fixes #1386