refactor: changed how `--force-type` works
Before the value of this parameter would be passed down so `Source`
could change the way it treats the code it parses. But now we simply
do what we do for extension-less files: we make a temporary copy
with a more correct name and use that instead of the original.
This simplifies how the code works because it "fixes" the issue at the
start instead of code "deeper down" needing to take this into account.