refactor: introduced `TagReader`
Extracted code from `Source` that would read the contents of the file
and find all the tag lines (the ones starting with `//`) and parse the
information contained in them. This is done so the code can be re-used
when reading JBang project files. This will allow us to treat those
files as actual project definitions instead of the current hack where
we treat them as source files (that must not contain actual code).