refactor: build.jbang files are now proper build files
We're not treating `build.jbang` files as a kind of `Source` anymore
which means we can change it's format when compared to `Source` files.
In this case a `build.jbang` file still supports all the tags that a
`Source` file supports but without the `//` prefix. It also isn't a
source file anymore so you can't put any code in the file.