Refactor RunContext into ProjectBuilder and other stuff (#1469)
* refactor: limit usage of `RunContext`

Don't use it for information that is already available in `Code` or can
be deduced in other ways. Introduced `NativeCmdGenerator`.

* refactor: Simplified `Code` hierarchy

Before we had the `Code` interface and 2 implementing classes: `Project`
and `Jar`, but there were a number of complex interactions between them
with lots of shared information the had to be kept in sync. Now there's
only `Project`.

* refactor: `RunContext` isn't passed around anymore

`RunContext` is now only used at the start for setting up a `Project`.
This is a first step towards turning it into a simple builder.

* feat: `--native` now works for jars

At least JBang will try to run the native image builder on the jar and
will run the binary if it was succesful.

* refactor: introduced the concept of `BuildSteps`

Split the larger `Java|Groovy|KotlinBuilder` classes up into smaller
`BuildStep` classes, each handling a very well-defined subset of the
build process.

* feat: `info tools` now returns path to native image (if exists)

* feat: Added `--build-dir` option

This tells JBang where to build the application. Normally JBang will
build in its special cache folder in the user's home directory. This
can be overridden but will a) cause everything to be cached in a
different place and b) the cached items can have strange names.
This option tells JBang to only store its build artifacts in the folder
specified by the user and it will do so using "proper" names.

There's also an additional, somewhat hidden, feature where you can tell
JBang to _not_ delete the class files that were generated during the
compilation build step. Simply use `-Djbang.build.keepclasses=true`.

* refactor: Agents are now part of the `Project`

* refactor: renamed some classes

`ProjectBuilder` is now called `AppBuilder` while `RunContext` is now
`ProjectBuilder`.

* refactor: changed how `--force-type` works

Before the value of this parameter would be passed down so `Source`
could change the way it treats the code it parses. But now we simply
do what we do for extension-less files: we make a temporary copy
with a more correct name and use that instead of the original.
This simplifies how the code works because it "fixes" the issue at the
start instead of code "deeper down" needing to take this into account.

* refactor: made `ProjectBuilder.createSource()` private

This removed the last public instance method from the class that has
noting to do with building a `Project`. It's now a true builder.
Also not passing propeties anymore to the `CmdGenerator` because they
can be obtained from `Project`.
Some minor cleanup.

* refactor: introduced `TagReader`

Extracted code from `Source` that would read the contents of the file
and find all the tag lines (the ones starting with `//`) and parse the
information contained in them. This is done so the code can be re-used
when reading JBang project files. This will allow us to treat those
files as actual project definitions instead of the current hack where
we treat them as source files (that must not contain actual code).

* refactor: extracted out "@Grab" annotation handling

This is now part of `TagReader.Extended` so we can have a base reader
that only deals with //-lines.

* refactor: build.jbang files are now proper build files

We're not treating `build.jbang` files as a kind of `Source` anymore
which means we can change it's format when compared to `Source` files.
In this case a `build.jbang` file still supports all the tags that a
`Source` file supports but without the `//` prefix. It also isn't a
source file anymore so you can't put any code in the file.