fix: we don't blindly import JAR metadata anymore
We added a way to skip importing metadata into a Project if a JAR was
available locally, but that was still causing problems in certain
situations. For example if you would compile with Java 19 and then
uninstall Java 19 leaving an earlier version, let's say Java 11, jbang
would try to download Java 19 just to run the code, even though nothing
in the source indicated that was necessary. In that case it's much more
convenient to simply recompile the code with Java 11. The solution now
is to not import the metadata into the Project anymore, but to make it
an explicit action of reading the information and let the code that does
the reading decide what to do with it.