refactor: Gist handling now uses core code for downloads
Before it would do it's own connection handling and did not honor
`--offline` and would do no caching. Now it uses the same code as all
other parts of the code that need to retrieve information over HTTP.
As an added benefit the GITHUB_TOKEN handling has now been made part of
the core network code so any access to GitHub can now use it.