fix: fixed minor GAV resolve ordering problem
If the resource to run/build/info is a GAV Jbang will resolve twice,
this is still something that should be fixed but for now that's just
how it works. It was considered okay because we cache resolve results
anyway. But it turns out there was a slight difference in how the
Maven query was ordered in both cases resulting in different cache hits.
This is now fixed.