feat: added support for modules
There is now a new source tag `//MODULE <name>` which signifies that
the code should be compiled as a module with the given name.
The dependencies defined are automatically added as required modules.
This of course requires at least Java 9 so the minimal version for
scripts with the `//MODULE` tag is automatically set to 9.