refactor: extracted build state from `Project`
The `Project` class now does not contain any build-related code anymore,
this has all been moved to `BuildContext`. So a `Builder` can be given
a `Project` and a `BuildContext` where the former represents all the
inputs that go into the build and the latter represents all the outputs
(and any temporary/intermediate state).