refactor: introduced `CmdGeneratorBuilder`
This is similar to what we already have with `ProjectBuilder` and it
used to pass user options to the generator (which is for running code)
that should not have been in `Project` (which is for building code).
Also changed the builder factory methods a bit.
Another thing this allows us to do is to have a notion of options that
can have different values for building and running. One example of that
is the `mainClass`. We want to be able to set the default main class for
building and have it create a JAR that knows which class to run. But we
also want to be able to override that at run time. So now both the
`ProjectBuilder` and the `CmdGeneratorBuilder` have a `mainClass`
attribute, something that wasn't possible before.