feat: added support for modules (#1567)
* feat: added support for modules

There is now a new source tag `//MODULE <name>` which signifies that
the code should be compiled as a module with the given name.
The dependencies defined are automatically added as required modules.
This of course requires at least Java 9 so the minimal version for
scripts with the `//MODULE` tag is automatically set to 9.

* feat: added support for `//MAIN`

The main class for a script can now be explicitly set by using the new
`//MAIN` tag. This overrides the default detection that Jbang  performs
and is assumed to be correct, no validation is done.

* refactor: introduced `CmdGeneratorBuilder`

This is similar to what we already have with `ProjectBuilder` and it
used to pass user options to the generator (which is for running code)
that should not have been in `Project` (which is for building code).
Also changed the builder factory methods a bit.
Another thing this allows us to do is to have a notion of options that
can have different values for building and running. One example of that
is the `mainClass`. We want to be able to set the default main class for
building and have it create a JAR that knows which class to run. But we
also want to be able to override that at run time. So now both the
`ProjectBuilder` and the `CmdGeneratorBuilder` have a `mainClass`
attribute, something that wasn't possible before.

* chore: show more information on HTTP errors

* docs: added some docs for `//MODULE` and `//MAIN`

* fix: fixed CDS support

Turns out CDS was broken for quite some time now. We didn't really
notice because Java silently ignores any issues with missing JSA files.
We now also print out a message when trying to enable CDS on Java <13.

* chore: merge changes and marking modules experimental

* Apply suggestions from code review

---------

Co-authored-by: Max Rydahl Andersen <max.andersen@gmail.com>