refactor: minor rollback of changes to `resolveInJavaHome()`
Earlier changes were moving in the direction of passing a `Jdk` object
instead of a version string. But that is actually not future save,
where there might be a point where we would want to be able to say:
"okay, I've used 'javac' from Jdk X to compile this and now I need
'native-image', but it's not in this Jdk X. Is there perhaps a newer
Jdk I could use?". If we pass a `Jdk` object we would be stuck at the
current version.