fix: `CmdGeneratorBuilder` now has its own `moduleName`
This is similar to `mainClass` which also exists both in the `Project`
and in the `CmdGenerator`. And for the same reason: we want to be able
to use a different value at runtime than that at buid time.
But unlike `mainClass` the value for `moduleName` is not always copied
as the default value from `Project` to `CmdGenerator`; we don't always
want to run JARs as modules by default, for example, but instead we'd
like the user to state their intentions explicitely by passing
`--module` to the `run` command.