feat: `//MODULE` is now more usable
JBang will now automatically add all system modules as required and
will open the script's package to everyone. This is not ideal because it
breaks all encapsulation, but it results in code that will work in most
cases. If you want better control you'll have to fall back to using a
separate `module-info.java` file.