test: fixed tests not always working right
First of all having a `@BeforeEach` with the same method name in a base
class and a sub class can sometimes cause problems, especially when you
try to run single tests.
And the second problem would normally occur only on Windows where the
TEMP folder is actually in the same filesystem tree as the user files,
meaning that when we created a temp folder strcture for the tests the
code that looks for aliases and catalogs etc would still return actual
user files, ruining the tests.