fix: `java-env X` will now use most specific path
Before if you would pass it a version that happened to be the default
version it would return the path to the "Current JDK", which can
change using `default`, so it's not really the most specific path.
Now it returns the path to the actual version the user asked for.