fix: no longer rebuilding source dependencies
Source dependencies would always be rebuilt whenever the parent script
was built. This was because they were really built a part of (meaning
inside of) the parent's build folder, not as an independent artifact.
This is now no longer the case and building a script as part of a
dependency or standalone will now result in the same artifact.