fix: Now showing integration errors when verbose is on (#1630)
* fix: Now showing integration errors when verbose is on

Fixes #1499

* fix: External integrations will now work in IDE

Before the class loader would not be set up correctly
because it expects to be running from an uber JAR with
all the dependencies available. Now we specifically
add the path to the Gson library in those cases.
The user won't notice but it makes life easier for
developers.