feat: added import feature to catalog refs
Catalog refs inside a catalog can now have an `import` field that when
set to `true` will make the catalog's items available to the user
without then having to explicitely mention the catalog. So for example
`jbang run foo@bar` becomes `jbang run foo` if the catalog `bar` is
defined in any of the catalogs in the user's scope on the file system
and it has the `import` field set to `true`.
This also adds the `--import` flag to `catalog add`.