feat: liunxdistro jdk provider
adds `linuxdistro` provider which for now searches `/usr/lib/jvm` for jdks.

It is not enabled by default as it is not possible to safely determine the version and build of the jdk
present or needed to be installed if mising.

Meaning to get it user need to select it `--jdk-provider=linuxdistro` or just enable all providers `--jdk-provider=all`
on command line or run `jbang config set jdk-provider linuxdistro` to make it default for all jbang invocations.