fix: move to use junit+allure over karate
I love Karate and I'll always be a fan of it but
as time passed I realized it is a big dependency to have
with all the javascript/nashorn compabilitie for something
that for us is just calling out to a shell and does string comparisons.

This starts removing need for using Karate and instead uses plain
JUnit as gradle integration tests + using Allure to collect intermediate step results.

That last part was/is a key feature of karate where we can see all the steps
before and after a reported failure.

To run the build and see report try out `just itestreport`