# This is a combination of 2 commits. # This is the 1st commit message:
fix: no exceptions on missing sibling (re)sources

We used to throw exceptions when sibling sources or resources were
missing or if an error occurred while resolving them. But this makes it
difficult to use the `jbang info` command on a source that has missing
siblings. So now we no longer throw exceptions, but use the special
`UnresolvableResourceRef` to indicate that the resource could not be
resolved. This allows the `jbang info` command to still work and show
information about the source, even if some of its siblings are missing.

Fixes #2069

# The commit message #2 will be skipped:

# wip