fix: `JBANG_JDK_VENDOR` can again be used (#2113)
The latest version of devkitman supports setting the "distro" of the JDK
to install, which allows us to use the `JBANG_JDK_VENDOR` environment
variable again to select the JDK vendor. Which was a feature that was
missing since the switch to devkitman. (In devkitman the term "distro"
was chosen instead of "vendor".)