feat: added `jdk exec` command (#2221)
* feat: added `jdk exec` command

Also introduced single letter aliases for several of the jdk subcommands.

* fix: made executing interactive commands work

We no longer execute the commands directly from Java but return them
back to the shell to be executed.

* fix: now also make things work for PowerShell

This was a bit more involved because PowerShell has no way to make
environments contextual. So you have to remember the old values of
variables you change and restore them at the end of the script.